_FASE1/apis/completo_fast_api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
SQLAlchemy==2.0.30
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
bcrypt==4.0.1

